+++ pdi/src/expression/impl/mapping.h
@@ -53,8 +53,6 @@ struct PDI_NO_EXPORT Expression::Impl::Mapping : public Expression::Impl {
 	
 	Ref to_ref(Context& ctx) const override;
 	
-	Ref to_ref(Context& ctx, const Datatype& type) const override;
-	
 	size_t copy_value(Context& ctx, void* buffer, const Datatype& type) const override;
 };